requirements.txt
algoliasearch-django
